# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: rez_agent API
  description: |
    REST API for the rez_agent event-driven messaging system.

    This API provides endpoints for managing messages, viewing metrics, and handling user authentication.
    All endpoints (except /health and /auth/*) require JWT authentication via AWS Cognito.

    ## Authentication
    - Obtain JWT token via OAuth 2.0 flow (POST /api/auth/login)
    - Include token in Authorization header: `Authorization: Bearer <token>`
    - Tokens expire after 1 hour (refresh via /api/auth/refresh)

    ## Environments
    - dev: https://api-dev.rez-agent.example.com
    - stage: https://api-stage.rez-agent.example.com
    - prod: https://api.rez-agent.example.com
  version: 1.0.0
  contact:
    name: rez_agent Team
    email: support@example.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api-dev.rez-agent.example.com
    description: Development environment
  - url: https://api-stage.rez-agent.example.com
    description: Staging environment
  - url: https://api.rez-agent.example.com
    description: Production environment
...
security:
  - BearerAuth: []
paths:
  /api/messages:
    get:
      tags:
        - messages
      summary: List messages
      description: |
        Retrieve a paginated list of messages with optional filtering.

        **Access Control**: Requires valid JWT token.

        **Query Parameters**:
        - Filter by stage, status, message_type, created_by
        - Date range filtering via created_after/created_before
        - Cursor-based pagination via cursor parameter
        - Results sorted by created_date (descending by default)

        **Performance**: Queries DynamoDB GSI (stage-created_date-index or status-created_date-index)
      operationId: listMessages
      parameters:
        - name: stage
          in: query
          description: Filter by environment stage
          required: false
          schema:
            type: string
            enum: [dev, stage, prod]
            example: dev
        - name: status
          in: query
          description: Filter by message status
          required: false
          schema:
            type: string
            enum: [created, queued, processing, completed, failed]
            example: completed
        - name: message_type
          in: query
          description: Filter by message type
          required: false
          schema:
            type: string
            example: scheduled_hello
        - name: created_by
          in: query
          description: Filter by creator (e.g., "scheduler", "user:email@example.com")
          required: false
          schema:
            type: string
            example: scheduler
        - name: created_after
          in: query
          description: Filter messages created after this timestamp (ISO 8601)
          required: false
          schema:
            type: string
            format: date-time
            example: "2025-10-21T00:00:00Z"
        - name: created_before
          in: query
          description: Filter messages created before this timestamp (ISO 8601)
          required: false
          schema:
            type: string
            format: date-time
            example: "2025-10-21T23:59:59Z"
        - name: limit
          in: query
          description: Maximum number of results to return (default 50, max 100)
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
            example: 50
        - name: cursor
          in: query
          description: Pagination cursor from previous response (base64-encoded LastEvaluatedKey)
          required: false
          schema:
            type: string
            example: eyJtZXNzYWdlX2lkIjogInV1aWQifQ==
      responses:
        '200':
          description: Successful response with message list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageListResponse'
              example:
                data:
                  - message_id: 550e8400-e29b-41d4-a716-446655440000
                    created_date: "2025-10-21T14:30:00Z"
                    created_by: scheduler
                    stage: dev
                    message_type: scheduled_hello
                    status: completed
                    payload:
                      text: hello world
                      priority: default
                      tags: [scheduled, daily]
                    updated_date: "2025-10-21T14:30:15Z"
                    processing_started_at: "2025-10-21T14:30:05Z"
                    processing_completed_at: "2025-10-21T14:30:15Z"
                    retry_count: 0
                pagination:
                  limit: 50
                  cursor: eyJtZXNzYWdlX2lkIjogIm5leHQtdXVpZCJ9
                  has_more: true
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
      security:
        - BearerAuth: []

    post:
      tags:
        - messages
      summary: Create a new message
      description: |
        Manually create a new message and publish to the processing pipeline.

        **Access Control**: Requires valid JWT token. Creator is automatically set to authenticated user.

        **Flow**:
        1. Validate request payload
        2. Create message record in DynamoDB (status="created")
        3. Publish message event to SNS topic
        4. Return 202 Accepted with message_id
        5. Async processing continues via SQS → Lambda

        **Idempotency**: Not idempotent. Each POST creates a new message with unique message_id.
      operationId: createMessage
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateMessageRequest'
            example:
              stage: dev
              message_type: manual_notification
              payload:
                text: Manual test message
                title: Test Notification
                priority: high
                tags: [manual, test]
      responses:
        '202':
          description: Message accepted for processing
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateMessageResponse'
              example:
                message_id: 660e9500-f39c-52e5-b827-557766551111
                status: created
                created_date: "2025-10-21T15:00:00Z"
                message: Message accepted for processing
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
      security:
        - BearerAuth: []
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT token issued by AWS Cognito.

        Include in Authorization header: `Authorization: Bearer <token>`

        Obtain token via POST /api/auth/login → POST /api/auth/callback flow.
  schemas:
    Message:
      type: object
      required:
        - message_id
        - created_date
        - created_by
        - stage
        - message_type
        - status
        - payload
        - updated_date
        - correlation_id
      properties:
        message_id:
          type: string
          format: uuid
          description: Unique message identifier
          example: 550e8400-e29b-41d4-a716-446655440000
        created_date:
          type: string
          format: date-time
          description: Message creation timestamp (ISO 8601)
          example: "2025-10-21T14:30:00Z"
        created_by:
          type: string
          description: Creator identifier (scheduler, user email, system)
          example: scheduler
        stage:
          type: string
          enum: [dev, stage, prod]
          description: Environment stage
          example: dev
        message_type:
          type: string
          description: Type of message/job
          example: scheduled_hello
        status:
          type: string
          enum: [created, queued, processing, completed, failed]
          description: Message lifecycle status
          example: completed
        payload:
          type: object
          description: Message content (varies by message_type)
          additionalProperties: true
          example:
            text: hello world
            priority: default
            tags: [scheduled, daily]
        updated_date:
          type: string
          format: date-time
          description: Last update timestamp
          example: "2025-10-21T14:30:15Z"
        processing_started_at:
          type: string
          format: date-time
          description: When processing began
          nullable: true
          example: "2025-10-21T14:30:05Z"
        processing_completed_at:
          type: string
          format: date-time
          description: When processing finished
          nullable: true
          example: "2025-10-21T14:30:15Z"
        error_message:
          type: string
          description: Error details if status=failed
          nullable: true
          example: ntfy.sh API returned 503 Service Unavailable
        retry_count:
          type: integer
          description: Number of retry attempts
          default: 0
          example: 0
        notification_id:
          type: string
          description: External notification identifier (e.g., ntfy.sh ID)
          nullable: true
          example: ntfy_abc123
        correlation_id:
          type: string
          format: uuid
          description: Distributed tracing correlation ID (X-Ray)
          example: 660e9500-f39c-52e5-b827-557766551111
    CreateMessageRequest:
      type: object
      required:
        - stage
        - message_type
        - payload
      properties:
        stage:
          type: string
          enum: [dev, stage, prod]
          description: Target environment stage
          example: dev
        message_type:
          type: string
          description: Type of message to create
          example: manual_notification
        payload:
          type: object
          description: Message content (schema varies by message_type)
          additionalProperties: true
          example:
            text: Manual test message
            priority: high
            tags: [manual, test]
    CreateMessageResponse:
      type: object
      required:
        - message_id
        - status
        - created_date
        - message
      properties:
        message_id:
          type: string
          format: uuid
          description: Unique identifier for created message
          example: 660e9500-f39c-52e5-b827-557766551111
        status:
          type: string
          enum: [created]
          description: Initial message status
          example: created
        created_date:
          type: string
          format: date-time
          description: Message creation timestamp
          example: "2025-10-21T15:00:00Z"
        message:
          type: string
          description: Success message
          example: Message accepted for processing
    MessageListResponse:
      type: object
      required:
        - data
        - pagination
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Message'
          description: List of messages
        pagination:
          type: object
          required:
            - limit
            - has_more
          properties:
            limit:
              type: integer
              description: Number of results per page
              example: 50
            cursor:
              type: string
              description: Cursor for next page (base64-encoded DynamoDB LastEvaluatedKey)
              nullable: true
              example: eyJtZXNzYWdlX2lkIjogIm5leHQtdXVpZCJ9
            has_more:
              type: boolean
              description: Whether more results are available
              example: true
...
    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: string
              description: Machine-readable error code
              example: VALIDATION_ERROR
            message:
              type: string
              description: Human-readable error message
              example: Invalid request parameters
            details:
              type: object
              description: Additional error context
              additionalProperties: true
              example:
                field: stage
                reason: must be one of [dev, stage, prod]
  responses:
    BadRequest:
      description: Invalid request parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: VALIDATION_ERROR
              message: Invalid request parameters
              details:
                field: stage
                reason: must be one of [dev, stage, prod]
    Unauthorized:
      description: Missing or invalid authentication token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: UNAUTHORIZED
              message: Invalid or expired JWT token
              details: {}
    TooManyRequests:
      description: Rate limit exceeded
      headers:
        Retry-After:
          description: Seconds until rate limit resets
          schema:
            type: integer
            example: 60
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: RATE_LIMIT_EXCEEDED
              message: Too many requests. Please try again in 60 seconds.
              details:
                retry_after: 60
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: INTERNAL_SERVER_ERROR
              message: An unexpected error occurred. Please try again later.
              details:
                correlation_id: 660e9500-f39c-52e5-b827-557766551111